bugfix in the moveRoot function
18th September 2013
-------------------
Fixed a problem with src/python/benLib/tree.py::moveRoot where not all the
tree branch lengths were being set and therefore defaulting to 0.001. It
was necessary to pass in the branch length of the internal node to the fn2
function. The function was also not dealing with the original root
correctly and some rearrangement of the brackets was needed.

This is an import of changes made on an undetermined version of Ortheus
used by Ensembl at the but not at the EBI
The changes have not been tested on this newer Ortheus codebase